play test 1

+++ src/com/github/dachhack/sprout/effects/Speck.java
@@ -384,7 +384,7 @@ public class Speck extends Image {
 				break;
 
 			case CHANGE:
-				am = (float) FloatMath.sqrt((p < 0.5f ? p : 1 - p) * 2);
+				am = FloatMath.sqrt((p < 0.5f ? p : 1 - p) * 2);
 				scale.y = (1 + p) * 0.5f;
 				scale.x = scale.y * FloatMath.cos(left * 15);
 				break;
