+++ src/com/github/dachhack/sprout/windows/WndTabbed.java
@@ -78,7 +78,7 @@ public class WndTabbed extends Window {
 		chrome.size(width + chrome.marginHor(), height + chrome.marginVer());
 
 		camera.resize((int) chrome.width,
-				(int) (chrome.marginTop() + height + tabHeight()));
+				chrome.marginTop() + height + tabHeight());
 		camera.x = (int) (Game.width - camera.screenWidth()) / 2;
 		camera.y = (int) (Game.height - camera.screenHeight()) / 2;
